Refactor: Move server input into separate file
